+++ 3rd/lua-hashsi/hashsi.h
@@ -6,7 +6,7 @@
 #define HASHSI_KEYLEN	50
 struct hashsi_node {
 	char key[HASHSI_KEYLEN+1];
-	int val;
+	int64_t val;
 	struct hashsi_node *next;
 };
 
@@ -23,6 +23,6 @@ void hashsi_init(struct hashsi *si, int max);
 void hashsi_clear(struct hashsi *si);
 struct hashsi_node *hashsi_lookup(struct hashsi *si, const char * key);
 void hashsi_remove(struct hashsi *si, const char * key);
-int hashsi_upsert(struct hashsi * si,const char * key,int val); 
+int hashsi_upsert(struct hashsi * si,const char * key,int64_t val); 
 int hashsi_full(struct hashsi *si);
 #endif
